fixup prune test case
