+++ crates/rooch-pruner/src/tests/mod.rs
@@ -1,5 +1,4 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
 
-mod pruner_utils;
 mod test_pruner;
